[refactor] 패키지 이동 및 인터페이스 이름 변경 ( #12 )
- user_coupon 패키지 -> coupon 패키지 : (UserCoupon, UserCouponRepository) 는 coupon 에 더 가까운 기능이라 생각하여 이동하였습니다.
