+++ src/main/java/com/coupon/issuecouponservice/repository/coupon/UserCouponRepository.java
@@ -1,0 +1,11 @@
+package com.coupon.issuecouponservice.repository.coupon;
+
+import com.coupon.issuecouponservice.domain.coupon.UserCoupon;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface UserCouponRepository extends JpaRepository<UserCoupon, Long> {
+    List<UserCoupon> findByUserId(Long userId);
+
+}
